Multiple fixes
Fixed a few leaks, and fixed the flame fractal just enough to make it look presentable. also added a zoom to it although it is very slow and only zooms towards the center.

+++ srcs/histogram_functions.c
@@ -12,11 +12,11 @@
 
 #include "../includes/fractol.h"
 
-long long	get_max_histogram(t_data *img)
+int	get_max_histogram(t_data *img)
 {
 	int			x;
 	int			y;
-	long long	t_color;
+	int			t_color;
 
 	y = 0;
 	t_color = 0;
